Refactor api.py: extract Memory + Chat-History + KG routes (#369)
Continues the api.py split. Stacks on top of #368.

- Memory CRUD + search (5 endpoints) → routes/memory.py
- /agents/{name}/chat-history full-text search (1 endpoint)
- KG graph + stats (2 endpoints)

Total: 8 endpoints moved.

`_get_memory_store` and `_reflection_to_dict` were used only by these
routes — both move into the new module so it's fully self-contained
for memory operations. The memory store is opened per-request from
`{working_dir}/data/memory.db` so there is no shared state to wire
through.

`_collect_agent_session_ids` and `_resolve_agent_history` stay in
api.py — they are also used by `dream_runner` and exposed via
`app.state.agent_history_resolver` — and are passed in via
set_dependencies as callables.

api.py: 8,028 → 7,760 lines (-268, -3.3%)
Cumulative since main: 11,138 → 7,760 (-3,378 lines, -30.3%)

Tests: 1542 passed, 1 skipped (federation tests need nacl, pre-existing).
